more robust check script
This is the check script from the jacques-clouseau repository, which was itself
modified from the original check script in this repo.

It avoids attempting to run shellcheck on something that isn't a regular file,
such as a directory (which wouldn't make sense) or symlink (which wouldn't
usually be desired, since the point of the script is to check the files in a
repository).

It also extracts the kludge to a named function.